Move terminalTab to browser layer
