+++ src/vs/workbench/parts/terminal/browser/terminalTab.ts
